refactor: add services package
