+++ app/src/main/java/com/example/eventapp/ui/profiles/ProfileEditFragment.java
@@ -35,8 +35,8 @@ import com.example.eventapp.R;
 import com.example.eventapp.databinding.FragmentProfileEditBinding;
 import com.example.eventapp.models.Facility;
 import com.example.eventapp.models.User;
-import com.example.eventapp.photos.PhotoPicker;
-import com.example.eventapp.photos.PhotoManager;
+import com.example.eventapp.services.photos.PhotoPicker;
+import com.example.eventapp.services.photos.PhotoManager;
 import com.example.eventapp.viewmodels.ProfileViewModel;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
